requirements.txt
requests
Pillow
flask

# rgbmatrix is NOT installed via pip.
# Install from source on the Pi:
#   git clone https://github.com/hzeller/rpi-rgb-led-matrix.git
#   cd rpi-rgb-led-matrix/bindings/python
#   make build-python PYTHON=$(which python3)
#   sudo make install-python PYTHON=$(which python3)
